Move hook to hook folder
