+++ libs/domains/variables/feature/src/index.ts
@@ -5,6 +5,7 @@ export * from './lib/hooks/use-delete-variable/use-delete-variable'
 export * from './lib/hooks/use-edit-variable/use-edit-variable'
 export * from './lib/hooks/use-import-variables/use-import-variables'
 export * from './lib/hooks/use-variables/use-variables'
+export * from './lib/hooks/use-variables-secret-managers/use-variables-secret-managers'
 export * from './lib/output-variables/output-variables'
 export * from './lib/variables-action-toolbar/variables-action-toolbar'
 export * from './lib/variable-list/variable-list'
@@ -15,4 +16,3 @@ export * from './lib/field-variable-suggestion/field-variable-suggestion'
 export * from './lib/dropdown-variable/dropdown-variable'
 export * from './lib/import-environment-variable-modal/import-environment-variable-modal'
 export * from './lib/external-secrets/add-secret-modal/add-secret-modal'
-export * from './lib/external-secrets/use-variables-secret-managers'
